+++ foxes/engines/futures.py
@@ -72,6 +72,12 @@ class ProcessEngine(ThreadsEngine):
     :group: engines
 
     """
+    def __init__(self, *args, **kwargs):
+        """
+        Constructor
+        """
+        super().__init__(*args, **kwargs)
+        self.share_cstore = False
 
     def _create_pool(self):
         """Creates the pool"""