+++ foxes/output/flow_plots_2d/__init__.py
@@ -1,2 +1,3 @@
 from .flow_plots import FlowPlots2D as FlowPlots2D
 from .get_fig import get_fig as get_fig
+from .from_chunk import write_chunk_ani_xy as write_chunk_ani_xy